extract review confirmation modal component
