+++ src/app/[orgSlug]/study/[studyId]/review/proposal-review-view.tsx
@@ -1,7 +1,7 @@
 'use client'
 
-import { AppModal } from '@/components/modal'
 import { PageBreadcrumbs } from '@/components/page-breadcrumbs'
+import { ReviewConfirmationModal, REJECTION_WARNING } from '@/components/review-confirmation-modal'
 import { useProposalReviewMutation } from '@/hooks/use-proposal-review-mutation'
 import { useReviewDecision } from '@/hooks/use-review-decision'
 import { useReviewFeedback } from '@/hooks/use-review-feedback'
@@ -14,7 +14,7 @@ import { Box, Button, Group, Stack, Text, Title } from '@mantine/core'
 import { useDisclosure } from '@mantine/hooks'
 import { CaretLeftIcon } from '@phosphor-icons/react'
 import { useRouter } from 'next/navigation'
-import { useState, type FC, type ReactNode } from 'react'
+import { useState, type FC } from 'react'
 import type { ProposalFeedbackEntry } from '@/server/actions/study.actions'
 import { FeedbackAndNotesSection } from '@/components/study/feedback-and-notes'
 import { ProposalSection } from './proposal-section'
@@ -115,62 +115,8 @@ const ReviewActionsBar: FC<ReviewActionsBarProps> = ({ study, canSubmit, isPendi
     )
 }
 
-type ReviewConfirmationModalProps = {
-    isOpen: boolean
-    onClose: () => void
-    onConfirm: () => void
-    isPending: boolean
-    title: string
-    confirmLabel: string
-    variant?: 'default' | 'destructive'
-    warning?: ReactNode
-}
-
-export const REJECTION_WARNING = (
-    <Text size="md" fw={600} c="red.9">
-        Rejection: This is intended as a last resort due to major, unresolvable issues and will end this study. This
-        action cannot be undone.
-    </Text>
-)
-
-export const ReviewConfirmationModal: FC<ReviewConfirmationModalProps> = ({
-    isOpen,
-    onClose,
-    onConfirm,
-    isPending,
-    title,
-    confirmLabel,
-    variant = 'default',
-    warning,
-}) => {
-    const isDestructive = variant === 'destructive'
-    return (
-        <AppModal
-            isOpen={isOpen}
-            onClose={onClose}
-            title={title}
-            size={720}
-            closeOnClickOutside={!isPending}
-            closeOnEscape={!isPending}
-            withCloseButton={!isPending}
-        >
-            <Stack>
-                <Text size="md">
-                    Please confirm you are ready to submit your review. Further edits are not permitted once submitted.
-                </Text>
-                {warning}
-                <Group justify="flex-end">
-                    <Button variant="outline" onClick={onClose} disabled={isPending}>
-                        Cancel
-                    </Button>
-                    <Button color={isDestructive ? 'red' : undefined} onClick={onConfirm} loading={isPending}>
-                        {confirmLabel}
-                    </Button>
-                </Group>
-            </Stack>
-        </AppModal>
-    )
-}
+const CONFIRM_BODY =
+    'Please confirm you are ready to submit your review. Further edits are not permitted once submitted.'
 
 /**
  * Inner component that does all the hook work. Lives **inside**
@@ -252,7 +198,9 @@ function ProposalReviewViewContent({ orgSlug, study, priorEntries, reviewVersion
                 isPending={isPending}
                 title="Confirm review submission?"
                 confirmLabel="Yes, submit review"
-            />
+            >
+                <Text size="md">{CONFIRM_BODY}</Text>
+            </ReviewConfirmationModal>
             <ReviewConfirmationModal
                 isOpen={rejectOpen}
                 onClose={closeReject}
@@ -261,8 +209,10 @@ function ProposalReviewViewContent({ orgSlug, study, priorEntries, reviewVersion
                 title="Reject initial request"
                 confirmLabel="Reject initial request"
                 variant="destructive"
-                warning={REJECTION_WARNING}
-            />
+            >
+                <Text size="md">{CONFIRM_BODY}</Text>
+                {REJECTION_WARNING}
+            </ReviewConfirmationModal>
         </Box>
     )
 }
